Reorganise components folder
