+++ qbd-frontend/src/app/components/book_details/component.book_details.css
