+++ qbd-frontend/src/app/components/book_details/component.book_details.ts
@@ -2,8 +2,8 @@ import { Component, effect, Input, Output, EventEmitter } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { switchMap, filter } from 'rxjs/operators';
 import { Observable, of, tap } from 'rxjs';
-import { BookService } from '../services/book_service';
-import { Book } from '../models/book';
+import { BookService } from '../../services/book_service';
+import { Book } from '../../models/book';
 import { inject } from '@angular/core';
 import { CommonModule } from '@angular/common';
