+++ qbd-frontend/src/app/components/book_list/component.book_list.css
