+++ qbd-frontend/src/app/components/footer/component.footer.html
