+++ qbd-frontend/src/app/components/footer/component.footer.ts
