+++ qbd-frontend/src/app/components/header/component.header.ts
