+++ qbd-frontend/src/app/components/homepage/component.homepage.html
